backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
cohere==5.5.8
qdrant-client==1.9.0
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
pydantic==2.5.0
langchain==0.2.17
langchain-community==0.2.19
langchain-cohere==0.3.0
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.2